requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
supabase
python-dotenv
bcrypt

